+++ lib/rubocop/cop/style/method_call_with_args_parentheses.rb
@@ -156,7 +156,7 @@ module RuboCop
           return if node.implicit_call?
           return if super_call_without_arguments?(node)
           return if camel_case_method_call_without_arguments?(node)
-          return if eligible_for_parentheses_presence?(node)
+          return if legitimate_call_with_parentheses?(node)
 
           add_offense(node, location: node.loc.begin.join(node.loc.end))
         end
@@ -225,10 +225,11 @@ module RuboCop
           node.camel_case_method? && node.arguments.none?
         end
 
-        def eligible_for_parentheses_presence?(node)
+        def legitimate_call_with_parentheses?(node)
           call_in_literals?(node) ||
             call_with_ambiguous_arguments?(node) ||
             call_in_logical_operators?(node) ||
+            call_in_optional_arguments?(node) ||
             allowed_multiline_call_with_parentheses?(node) ||
             allowed_chained_call_with_parentheses?(node)
         end
@@ -247,17 +248,22 @@ module RuboCop
              node.parent.descendants.any?(&method(:logical_operator?)))
         end
 
+        def call_in_optional_arguments?(node)
+          node.parent && node.parent.optarg_type?
+        end
+
         def call_with_ambiguous_arguments?(node)
           call_with_braced_block?(node) ||
             call_as_argument_or_chain?(node) ||
             hash_literal_in_arguments?(node) ||
             node.descendants.any? do |n|
-              splat?(n) || ternary_if?(n) || logical_operator?(n)
+              ambigious_literal?(n) || logical_operator?(n) ||
+                call_with_braced_block?(n)
             end
         end
 
         def call_with_braced_block?(node)
-          node.block_node && node.block_node.braces?
+          node.send_type? && node.block_node && node.block_node.braces?
         end
 
         def call_as_argument_or_chain?(node)
@@ -285,6 +291,10 @@ module RuboCop
             allowed_chained_call_with_parentheses?(previous)
         end
 
+        def ambigious_literal?(node)
+          splat?(node) || ternary_if?(node) || regexp_slash_literal?(node)
+        end
+
         def splat?(node)
           node.splat_type? || node.kwsplat_type? || node.block_pass_type?
         end
@@ -299,6 +309,10 @@ module RuboCop
 
         def hash_literal?(node)
           node.hash_type? && node.braces?
+        end
+
+        def regexp_slash_literal?(node)
+          node.regexp_type? && node.loc.begin.source == '/'
         end
       end
     end